+++ src/main/java/com/imveis/visita/Imoveis/repositories/ImovelRepository.java
@@ -3,6 +3,8 @@ package com.imveis.visita.Imoveis.repositories;
 import com.imveis.visita.Imoveis.entities.Imovel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImovelRepository extends JpaRepository<Imovel, Long> {
+import java.math.BigInteger;
+
+public interface ImovelRepository extends JpaRepository<Imovel, BigInteger> {
 
 }
